Updated
I was only able to make a minor change to the code by removing the explicit "String" definitions to the AircraftCamera enum. Trying to rework all of the code is apparently difficult bc, yes, setting a reading an AircraftCamera enum typed property is fine until I _need_ a String type, as in the AircraftCameraButtonsView, where I'm reading a string from button presses and then setting that as the SCNNode's name property, which takes a String.